+++ src/collective/contact/core/tests/test_browser_utils.py
@@ -1,6 +1,6 @@
 # -*- coding: utf8 -*-
-from collective.contact.core import logged_actions
 from collective.contact.core.testing import FUNCTIONAL
+from collective.contact.core.testing import logged_actions
 from ecreall.helpers.testing.base import BaseTest
 from plone import api
 from plone.app.testing import setRoles